logou using firebase

+++ client/src/services/authService.js
@@ -1,11 +1,13 @@
 import { auth, firestore } from "./firebaseConfig";
 
-import { collection, setDoc, doc } from "firebase/firestore";
+import { collection, setDoc, doc, getDoc } from "firebase/firestore";
 import {
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
+  signOut,
 } from "firebase/auth";
 
+//store sign up user to firebase store:
 export const addUserToFirestore = async (userData) => {
   try {
     //created collection
@@ -24,21 +26,20 @@ export const addUserToFirestore = async (userData) => {
   }
 };
 
+//signup user
 export const signUpWithEmailPassword = async (email, password, displayName) => {
-  // console.log(email, password, displayName);
   try {
     const userCredential = await createUserWithEmailAndPassword(
       auth,
       email,
-      password,
-      displayName
+      password
     );
-    const user = userCredential.user;
+    const user = userCredential?.user;
 
     // Add additional user information to Firestore database
     const storeCredentials = {
-      uid: user.uid,
-      email: user.email,
+      uid: user?.uid,
+      email: user?.email,
       name: displayName,
       createdAt: new Date().toLocaleString(),
     };
@@ -59,20 +60,40 @@ export const loginWithEmailPassword = async (email, password) => {
     const user = res.user;
     console.log({ user });
 
-    //const res=await getUserDetails(users,user.uid);
+    // Fetch additional user details from Firestore based on the UID
+    const userDocRef = doc(firestore, "users", user.uid);
+    const userSnapshot = await getDoc(userDocRef);
 
-    const userDetails = {
-      token: user.accessToken,
-      uid: user.uid,
-      email: user.email,
-      isEmailVerified: user.emailVerified,
-      name: user.displayName,
-    };
-    // Storing user details in localStorage
-    localStorage.setItem("userDetails", JSON.stringify(userDetails));
+    if (userSnapshot.exists()) {
+      const userDetails = {
+        token: user.accessToken,
+        uid: user.uid,
+        email: user.email,
+        isEmailVerified: user.emailVerified,
+        name: userSnapshot.data().name,
+      };
 
-    return { success: true, user: userDetails };
+      // Storing user details in localStorage
+      localStorage.setItem("userDetails", JSON.stringify(userDetails));
+      return { success: true, user: userDetails };
+    } else {
+      return { success: false, error: "User details not found" };
+    }
   } catch (error) {
     return { success: false, error: error.message };
+  }
+};
+
+export const logOutUserWithEmailPasword = async () => {
+  try {
+    console.log("logout request received");
+    await signOut(auth);
+
+    console.log("logout req. processed");
+    localStorage.removeItem("userDetails");
+    return { success: true };
+  } catch (error) {
+    console.error("Error logging out:", error);
+    return { success: false, error: "Error logging out" };
   }
 };
